feat: Aprimore o gerenciamento de itens com funcionalidade de registro de perdas e atualize os tipos de ID para Long

+++ src/main/java/com/codenine/managementservice/dto/ItemFilterCriteria.java
@@ -2,10 +2,10 @@ package com.codenine.managementservice.dto;
 
 
 public record ItemFilterCriteria(
-    Integer supplierId,
-    Integer sectionId,
-    Integer typeItemId,
-    Integer lastUserId,
+    Long supplierId,
+    Long sectionId,
+    Long typeItemId,
+    Long lastUserId,
     Boolean isActive,
-    Integer itemId
+    Long itemId
 ) {}
